% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{acvToeplitz}
\alias{acvToeplitz}
\title{Build an nd times nd Block Toeplitz matrix from the
(d times d) autocovariances gamma(0),...,gamma(n-1) in beyondWhittle(misc.cpp)
called acvBlockMatrix()}
\usage{
acvToeplitz(acv)
}
\description{
Build an nd times nd Block Toeplitz matrix from the
(d times d) autocovariances gamma(0),...,gamma(n-1) in beyondWhittle(misc.cpp)
called acvBlockMatrix()
}
\keyword{internal}
