% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_multivariate_util.R
\name{Adj}
\alias{Adj}
\title{Adjoint of complex matrix. This Adj() in beyondWhittle(misc.R).}
\usage{
Adj(m)
}
\description{
Adjoint of complex matrix. This Adj() in beyondWhittle(misc.R).
}
\keyword{internal}
