% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{beta_fun_AGamma_process_cube}
\alias{beta_fun_AGamma_process_cube}
\title{Beta function, see Definition 3.3.
This is AGammaProcessPrior::beta() in beyondWhittle(a_gamma_process_prior.cpp)}
\usage{
beta_fun_AGamma_process_cube(U_, Sigma_)
}
\description{
Beta function, see Definition 3.3.
This is AGammaProcessPrior::beta() in beyondWhittle(a_gamma_process_prior.cpp)
}
\keyword{internal}
