% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_multivariate_util.R
\name{center}
\alias{center}
\title{Mean center a numerical vector. The is center() in beyondWhittle(misc.R).}
\usage{
center(x, ...)
}
\description{
Mean center a numerical vector. The is center() in beyondWhittle(misc.R).
}
\keyword{internal}
