% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chol_cube}
\alias{chol_cube}
\title{Cholesky docomposition representation of a matrix array. See Remark 5.2}
\usage{
chol_cube(f_, excludeBoundary)
}
\description{
Cholesky docomposition representation of a matrix array. See Remark 5.2
}
\keyword{internal}
