% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mittelbach_util.R
\name{cholesky_I_l}
\alias{cholesky_I_l}
\title{Range intervals I_l, see (63) in Mittelbach et al.
This is unit_trace_I_l() in beyondWhittle(unit_trace.R)}
\usage{
cholesky_I_l(l)
}
\description{
Range intervals I_l, see (63) in Mittelbach et al.
This is unit_trace_I_l() in beyondWhittle(unit_trace.R)
}
\keyword{internal}
