% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cholesky_jacobianLogDeterminant}
\alias{cholesky_jacobianLogDeterminant}
\title{log determinant of Jacobian, see (5.12)
This is unit_trace_jacobian_log_determinant() in beyondWhittle(unit_trace.cpp)}
\usage{
cholesky_jacobianLogDeterminant(phi)
}
\description{
log determinant of Jacobian, see (5.12)
This is unit_trace_jacobian_log_determinant() in beyondWhittle(unit_trace.cpp)
}
\keyword{internal}
