% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mittelbach_util.R
\name{cholesky_log_cVec}
\alias{cholesky_log_cVec}
\title{Get log(c) vector, see (70) in Mittelbach et al.
Helping function for \code{unit_trace_runif}
This is unit_trace_log_c() in beyondWhittle(unit_trace.R)}
\usage{
cholesky_log_cVec(p, q)
}
\description{
Get log(c) vector, see (70) in Mittelbach et al.
Helping function for \code{unit_trace_runif}
This is unit_trace_log_c() in beyondWhittle(unit_trace.R)
}
\keyword{internal}
