% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mittelbach_util.R
\name{cholesky_log_dVec}
\alias{cholesky_log_dVec}
\title{Get log(d) vector, see (39) in Mittelbach et al, adjusted to complex case
Helping function for \code{unit_trace_runif}
This unit_trace_log_d() in beyondWhittle(unit_trace.R)}
\usage{
cholesky_log_dVec(p, q)
}
\description{
Get log(d) vector, see (39) in Mittelbach et al, adjusted to complex case
Helping function for \code{unit_trace_runif}
This unit_trace_log_d() in beyondWhittle(unit_trace.R)
}
\keyword{internal}
