% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mittelbach_util.R
\name{cholesky_log_f_l}
\alias{cholesky_log_f_l}
\title{Get log(f_l), see (66) in Mittelbach et al.
Helping function for \code{unit_trace_runif}
This is unit_trace_log_f_l() in beyondWhittle(unit_trace.R)}
\usage{
cholesky_log_f_l(phi, p, q, log_c, l)
}
\description{
Get log(f_l), see (66) in Mittelbach et al.
Helping function for \code{unit_trace_runif}
This is unit_trace_log_f_l() in beyondWhittle(unit_trace.R)
}
\keyword{internal}
