% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mittelbach_util.R
\name{cholesky_muVec}
\alias{cholesky_muVec}
\title{Get mu vector, see (36) in Mittelbach et al.
Helping function for \code{unit_trace_runif}
This is unit_trace_mu() in beyondWhittle(unit_trace.R)}
\usage{
cholesky_muVec(p, q)
}
\description{
Get mu vector, see (36) in Mittelbach et al.
Helping function for \code{unit_trace_runif}
This is unit_trace_mu() in beyondWhittle(unit_trace.R)
}
\keyword{internal}
