% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cholesky_pVec}
\alias{cholesky_pVec}
\title{Get p vector, see (67) in Mittelbach et al.
This is unit_trace_p() in beyondWhittle(unit_trace.cpp)}
\usage{
cholesky_pVec(d)
}
\description{
Get p vector, see (67) in Mittelbach et al.
This is unit_trace_p() in beyondWhittle(unit_trace.cpp)
}
\keyword{internal}
