% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cholesky_qVec}
\alias{cholesky_qVec}
\title{Get q vector, see (68) in Mittelbach et al.
This is unit_trace_q() in beyondWhittle(unit_trace.cpp)}
\usage{
cholesky_qVec(d)
}
\description{
Get q vector, see (68) in Mittelbach et al.
This is unit_trace_q() in beyondWhittle(unit_trace.cpp)
}
\keyword{internal}
