% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mittelbach_util.R
\name{cholesky_runif}
\alias{cholesky_runif}
\title{Draw uniformly from Hpd matrices with unit trace
This is unit_trace_runif() in beyondWhittle(unit_trace.R)}
\usage{
cholesky_runif(n, d, verbose = F)
}
\description{
Draw uniformly from Hpd matrices with unit trace
This is unit_trace_runif() in beyondWhittle(unit_trace.R)
}
\keyword{internal}
