% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mittelbach_util.R
\name{cholesky_sigma2Vec}
\alias{cholesky_sigma2Vec}
\title{Get sigma2 vector, see (37) in Mittelbach et al.
Helping function for \code{unit_trace_runif}
This is unit_trace_sigma2() in beyondWhittle(unit_trace.R)}
\usage{
cholesky_sigma2Vec(p, q)
}
\description{
Get sigma2 vector, see (37) in Mittelbach et al.
Helping function for \code{unit_trace_runif}
This is unit_trace_sigma2() in beyondWhittle(unit_trace.R)
}
\keyword{internal}
