% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mittelbach_util.R
\name{cholesky_UFromPhi}
\alias{cholesky_UFromPhi}
\title{Get U (Hpd with unit trace) matrix from
phi (hyperspherical coordinates) vector.
This is unit_trace_U_from_phi() in beyondWhittle(unit_trace.R)}
\usage{
cholesky_UFromPhi(phi)
}
\description{
Get U (Hpd with unit trace) matrix from
phi (hyperspherical coordinates) vector.
This is unit_trace_U_from_phi() in beyondWhittle(unit_trace.R)
}
\keyword{internal}
