% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cholesky_xFromPhi}
\alias{cholesky_xFromPhi}
\title{Get x from phi, see (62) in Mittelbach et al.
This is unit_trace_x_from_phi() in beyondWhittle(unit_trace.cpp)}
\usage{
cholesky_xFromPhi(phi)
}
\description{
Get x from phi, see (62) in Mittelbach et al.
This is unit_trace_x_from_phi() in beyondWhittle(unit_trace.cpp)
}
\keyword{internal}
