% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mittelbach_util.R
\name{cholesy_log_nuVec}
\alias{cholesy_log_nuVec}
\title{Get log(nu) vector, see (38) in Mittelbach et al.
Helping function for \code{unit_trace_runif}
This is unit_trace_nu() in beyondWhittle(unit_trace.R)}
\usage{
cholesy_log_nuVec(sigma2_vec, log_c_vec, log_d_vec)
}
\description{
Get log(nu) vector, see (38) in Mittelbach et al.
Helping function for \code{unit_trace_runif}
This is unit_trace_nu() in beyondWhittle(unit_trace.R)
}
\keyword{internal}
