% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cholseky_LFromx}
\alias{cholseky_LFromx}
\title{Get L (lower triangular Cholesky) from x
Called U^* in Mittelbach et al, see (60) there
This is unit_trace_L_from_x() in beyondWhittle(unit_trace.cpp)}
\usage{
cholseky_LFromx(x)
}
\description{
Get L (lower triangular Cholesky) from x
Called U^* in Mittelbach et al, see (60) there
This is unit_trace_L_from_x() in beyondWhittle(unit_trace.cpp)
}
\keyword{internal}
