% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_util.R
\name{coarsened_bernstein}
\alias{coarsened_bernstein}
\title{Construct coarsened Bernstein polynomial basis of degree l on omega
This is in beyondWhittle(bernstein_polynomials.R)}
\usage{
coarsened_bernstein(omega, l)
}
\arguments{
\item{omega}{numeric vector in [0,1] of evaluation points}

\item{l}{positive integer for the degree}
}
\description{
Construct coarsened Bernstein polynomial basis of degree l on omega
This is in beyondWhittle(bernstein_polynomials.R)
}
\details{
See Appendix E.1 in Ghosal/Van der Vaart, Fundamentals, (2017)
}
\keyword{internal}
