% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_util.R
\name{coarsened_bernstein_i}
\alias{coarsened_bernstein_i}
\title{Helping function for \code{coarsened_bernstein}
This is in beyondWhittle(bernstein_polynomials.R)}
\usage{
coarsened_bernstein_i(omega, l, i)
}
\description{
Helping function for \code{coarsened_bernstein}
This is in beyondWhittle(bernstein_polynomials.R)
}
\keyword{internal}
