% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{const_cube}
\alias{const_cube}
\title{Transform all matrices of the array into the same desirable matrix}
\usage{
const_cube(sigma, N)
}
\description{
Transform all matrices of the array into the same desirable matrix
}
