% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGammaProcess_util.R
\name{create_omega_fun_from_beta_density}
\alias{create_omega_fun_from_beta_density}
\title{Returns a function taking one argument -- scaled beta density for the integration. See Lemma 3.15}
\usage{
create_omega_fun_from_beta_density(C_alpha, g0.alpha, g0.beta)
}
\description{
Returns a function taking one argument -- scaled beta density for the integration. See Lemma 3.15
}
\keyword{internal}
