% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cube_from_NumericVector}
\alias{cube_from_NumericVector}
\title{I/O: Only use *within* Rcpp}
\usage{
cube_from_NumericVector(x)
}
\description{
I/O: Only use *within* Rcpp
}
\keyword{beyondWhittle(matrix_cube.cpp)}
\keyword{in}
\keyword{internal}
