% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cubeTimesVector}
\alias{cubeTimesVector}
\title{Get the product of U and r for the construction of the Gamma process. See (5.7)
This is inside of bernsteinGammaPsd::get_W() in beyondWhittle(bernstein_gamma_psd.cpp).}
\usage{
cubeTimesVector(U_, r)
}
\description{
Get the product of U and r for the construction of the Gamma process. See (5.7)
This is inside of bernsteinGammaPsd::get_W() in beyondWhittle(bernstein_gamma_psd.cpp).
}
\keyword{internal}
