% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cx_cube_from_ComplexVector}
\alias{cx_cube_from_ComplexVector}
\title{I/O: Only use *within* Rcpp in beyondWhittle(matrix_cube.cpp)}
\usage{
cx_cube_from_ComplexVector(x)
}
\description{
I/O: Only use *within* Rcpp in beyondWhittle(matrix_cube.cpp)
}
\keyword{internal}
