% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_util.R
\name{dbList}
\alias{dbList}
\title{Construct Bernstein polynomial basises of degree up to kmax on omega
This is in beyondWhittle(bernstein_polynomials.R)}
\usage{
dbList(
  n,
  kmax,
  normalized = F,
  bernstein_l = 0,
  bernstein_r = 1,
  coarsened = F
)
}
\arguments{
\item{n}{positive integer determining the number of the (equidistant) evaluation points in [0,1]}

\item{kmax}{positive integer for the largest degree}

\item{bernstein_l, bernstein_r}{left and right truncation}

\item{coarsened}{bool flag indicating whether coarsened or standard Bernstein polynomials are used}

\item{verbose}{debugging parameter}
}
\value{
A list of length kmax, where the k-th list element is a matrix containing the polynomial basis of degree k
}
\description{
Construct Bernstein polynomial basises of degree up to kmax on omega
This is in beyondWhittle(bernstein_polynomials.R)
}
\keyword{internal}
