% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{epsilon_var}
\alias{epsilon_var}
\title{Epsilon process (residuals) of VAR model in beyondWhittle(varma.cpp)}
\usage{
epsilon_var(zt, ar)
}
\description{
Epsilon process (residuals) of VAR model in beyondWhittle(varma.cpp)
}
\keyword{internal}
