% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_CFZ}
\alias{get_CFZ}
\title{Build the correction matrix in the frequency domain}
\usage{
get_CFZ(FZ, f_half_inv_, f_param_half_, excludeBoundary)
}
\description{
Build the correction matrix in the frequency domain
}
\keyword{internal}
