% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_CFZ_q}
\alias{get_CFZ_q}
\title{Build the correction matrix in the frequency domain based on the q-parametrization}
\usage{
get_CFZ_q(FZ, q_, f_param_half_, excludeBoundary)
}
\description{
Build the correction matrix in the frequency domain based on the q-parametrization
}
\keyword{internal}
