% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_CFZ_q_sq}
\alias{get_CFZ_q_sq}
\title{Build the correction matrix in the frequency domain based on the q-parametrization with square root of matrices}
\usage{
get_CFZ_q_sq(FZ, q_, f_param_half_, excludeBoundary)
}
\description{
Build the correction matrix in the frequency domain based on the q-parametrization with square root of matrices
}
\keyword{internal}
