% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_gamma_matrix_subordinator_core.R
\name{get_f_matrix}
\alias{get_f_matrix}
\title{Construct psd mixture. See (5.3). We do not use the smoothing Choleksy components
method, so delete that in the future. This is get_f_matrix() in beyondWhittle(gibbs_vnp_help.cpp)}
\usage{
get_f_matrix(U, r, Z, k, db.list, prior.cholesky)
}
\description{
Construct psd mixture. See (5.3). We do not use the smoothing Choleksy components
method, so delete that in the future. This is get_f_matrix() in beyondWhittle(gibbs_vnp_help.cpp)
}
\keyword{internal}
