% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_mix_rcpp}
\alias{get_mix_rcpp}
\title{Add W * b_{j,k} to the psd f (W being the Gamma process weight and b_{j,k}
being the j'the basis polynomial of degree k) See (5.3).
This is inside of bernsteinGammaPsd() in beyondWhittle(bernstein_gamma_psd.cpp),
in which the function is classified into two versions: one is the local
recomputation and one is the full recomputation}
\usage{
get_mix_rcpp(w_, densities)
}
\description{
Add W * b_{j,k} to the psd f (W being the Gamma process weight and b_{j,k}
being the j'the basis polynomial of degree k) See (5.3).
This is inside of bernsteinGammaPsd() in beyondWhittle(bernstein_gamma_psd.cpp),
in which the function is classified into two versions: one is the local
recomputation and one is the full recomputation
}
\keyword{internal}
