% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_w_rcpp}
\alias{get_w_rcpp}
\title{Get the Gamma process weight. See (5.7). This is inside of bernsteinGammaPsd::get_W()
in beyondWhittle(bernstein_gamma_psd.cpp).}
\usage{
get_w_rcpp(p_, Z, k)
}
\description{
Get the Gamma process weight. See (5.7). This is inside of bernsteinGammaPsd::get_W()
in beyondWhittle(bernstein_gamma_psd.cpp).
}
\keyword{internal}
