% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_gamma_matrix_subordinator.R
\name{gibbs_m_nuisance}
\alias{gibbs_m_nuisance}
\title{Title}
\usage{
gibbs_m_nuisance(data, mcmc_params, corrected, prior_params, model_params)
}
\description{
Title
}
\keyword{internal}
