% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_vnpc.R
\name{gibbs_vnpc}
\alias{gibbs_vnpc}
\title{Gibbs sampler for multivariate Bayesian inference with the nonparametrically corrected VAR likelihood}
\usage{
gibbs_vnpc(
  data,
  var.order,
  Ntotal,
  burnin,
  thin = 1,
  print_interval = 100,
  numerical_thresh = 1e-12,
  adaption.N = burnin,
  adaption.batchSize = 50,
  adaption.tar = 0.44,
  eta = ncol(data),
  omega_fun = create_omega_fun_from_beta_density(1, 1, 1),
  Sigma_fun = my_Sigma_fun,
  k.theta = 0.01,
  kmax = 100 * coars + 500 * (!coars),
  trunc_l = 0.1,
  trunc_r = 0.9,
  coars = F,
  L = max(20, length(data)^(1/3)),
  mu_beta = rep(1e-04, ncol(data) * ncol(data) * var.order),
  V_beta = diag(ncol(data) * ncol(data) * var.order) * 10000
)
}
\arguments{
\item{data}{numerical matrix}

\item{var.order}{VAR order for the parametric working model}

\item{Ntotal}{total number of iterations to run the Markov chain}

\item{burnin}{number of initial iterations to be discarded}

\item{thin}{thinning number (postprocessing)}

\item{print_interval}{number of iterations, after which a status is printed to console}

\item{numerical_thresh}{lower (numerical pointwise) bound for the eigenvalues of the spectral density}

\item{adaption.N}{total number of iterations, in which the proposal variances (of r, U and VAR coefficients) are adapted}

\item{adaption.batchSize}{batch size of proposal adaption}

\item{adaption.tar}{target acceptance rate for adapted parameters}

\item{eta}{AGamma process parameter, real number > ncol(data) - 1}

\item{omega_fun}{AGamma process parameter, positive constant}

\item{Sigma_fun}{AGamma process parameter, Hpd matrix}

\item{k.theta}{prior parameter for polynomial degree k (propto exp(-k.theta*k*log(k)))}

\item{kmax}{upper bound for polynomial degree of Bernstein-Dirichlet mixture (can be set to Inf, algorithm is faster with kmax<Inf due to pre-computation of basis functions, but values 500<kmax<Inf are very memory intensive)}

\item{trunc_l, trunc_r}{left and right truncation of Bernstein polynomial basis functions, 0<=trunc_l<trunc_r<=1}

\item{coars}{flag indicating whether coarsened or default bernstein polynomials are used (see Appendix E.1 in Ghosal and van der Vaart 2017)}

\item{L}{truncation parameter of Gamma process}

\item{mu_beta}{prior parameter for VAR coefficients, stacked numerical vector}

\item{V_beta}{prior parameter for VAR coefficients, Hpd matrix}
}
\description{
Obtain samples of the posterior of the multivariate corrected likelihood in conjuction with an Hpd AGamma process prior on the spectral density matrix
}
