% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_multivariate_util.R
\name{is_hpd}
\alias{is_hpd}
\title{Is the matrix a Hpd matrix? This is in beyondWhittle(misc.R)}
\usage{
is_hpd(A, tol = 1e-15)
}
\description{
Is the matrix a Hpd matrix? This is in beyondWhittle(misc.R)
}
\keyword{internal}
