% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mittelbach_util.R
\name{is_quadratic}
\alias{is_quadratic}
\title{Is l quadratic? in beyondWhittle(unit_trace.R)}
\usage{
is_quadratic(l, thresh = 1e-15)
}
\description{
Is l quadratic? in beyondWhittle(unit_trace.R)
}
\keyword{internal}
