% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lalphaStar_AGamma}
\alias{lalphaStar_AGamma}
\title{Log alpha star density function for the prior of AGamma process. See (3.26) in Meier (2018).
This is AGammaProcessPrior::lalpha() in beyondWhittle(a_gamma_process_prior.cpp).}
\usage{
lalphaStar_AGamma(U_, eta, Sigma_)
}
\description{
Log alpha star density function for the prior of AGamma process. See (3.26) in Meier (2018).
This is AGammaProcessPrior::lalpha() in beyondWhittle(a_gamma_process_prior.cpp).
}
