% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_multivariate_util.R
\name{llike_var_full}
\alias{llike_var_full}
\title{VAR(p) full likelihood
This is in beyondWhittle(varma.R)}
\usage{
llike_var_full(zt, ar, sigma)
}
\description{
VAR(p) full likelihood
This is in beyondWhittle(varma.R)
}
\keyword{internal}
