% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_multivariate_util.R
\name{llike_var_partial}
\alias{llike_var_partial}
\title{VAR(p) partial likelihood (unnormalized)
Note: Fine for fixed p, but not suited for model comparison
This is in beyondWhittle(varma.R)}
\usage{
llike_var_partial(zt, ar, sigma)
}
\description{
VAR(p) partial likelihood (unnormalized)
Note: Fine for fixed p, but not suited for model comparison
This is in beyondWhittle(varma.R)
}
\keyword{internal}
