% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_multivariate_util.R
\name{mdft}
\alias{mdft}
\title{Multivariate discrete (fast) Fourier Transform in beyondWhittle(fourier_transform.R)}
\usage{
mdft(Z, real = F)
}
\description{
Multivariate discrete (fast) Fourier Transform in beyondWhittle(fourier_transform.R)
}
\keyword{internal}
