% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_multivariate_util.R
\name{midft}
\alias{midft}
\title{Multivariate inverse discrete (fast) Fourier Transform in beyondWhittle(fourier_transform.R)}
\usage{
midft(FZ, real = F)
}
\description{
Multivariate inverse discrete (fast) Fourier Transform in beyondWhittle(fourier_transform.R)
}
\keyword{internal}
