% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_multivariate_util.R
\name{mpdgrm}
\alias{mpdgrm}
\title{Compute Periodgram matrix from (complex-valued) Fourier coefficients in beyondWhittle(fourier_transform.R)}
\usage{
mpdgrm(FZ)
}
\description{
Compute Periodgram matrix from (complex-valued) Fourier coefficients in beyondWhittle(fourier_transform.R)
}
\details{
see (4.7) in Meier (2018)
}
\keyword{internal}
