% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGammaProcess_util.R
\name{my_Sigma_fun}
\alias{my_Sigma_fun}
\title{Obtain the initial value of Sigma for AGamma process,
This is simplified in beyondWhittle}
\usage{
my_Sigma_fun(x)
}
\description{
Obtain the initial value of Sigma for AGamma process,
This is simplified in beyondWhittle
}
\keyword{internal}
