% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{numericalUnstable}
\alias{numericalUnstable}
\title{Is the conditional number too large? See Section 5.2.3 in Meier (2008)}
\usage{
numericalUnstable(f_, excludeBoundary, TOL = 1e-12)
}
\description{
Is the conditional number too large? See Section 5.2.3 in Meier (2008)
}
\keyword{internal}
