% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_multivariate_util.R, R/gibbs_util.R
\name{omegaFreq}
\alias{omegaFreq}
\title{Fourier frequencies rescaled on the unit interval
This is omegaFreq() in beyondWhittle(fourier_transform.R)}
\usage{
omegaFreq(n)

omegaFreq(n)
}
\description{
Fourier frequencies rescaled on the unit interval
This is omegaFreq() in beyondWhittle(fourier_transform.R)

Fourier frequencies rescaled on the unit interval in beyondWhittle(fourier_transform.R)
}
\keyword{internal}
