% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_multivariate_util.R
\name{phiFromBeta_normalInverseWishart}
\alias{phiFromBeta_normalInverseWishart}
\title{Convert vector parametrization (beta) to matrix-parametrization (phi),
the latter as e.g. used in MTS::VAR()$ar.
This is in beyondWhittle(varma.R) with the same name.}
\usage{
phiFromBeta_normalInverseWishart(beta, K, p)
}
\arguments{
\item{beta}{coefficient vector, of dimension K*d*d}

\item{K}{positive integer, vector dimensionality}

\item{p}{nonnegarive integer, VAR order}
}
\value{
K times K*p coefficient matrix
}
\description{
Convert vector parametrization (beta) to matrix-parametrization (phi),
the latter as e.g. used in MTS::VAR()$ar.
This is in beyondWhittle(varma.R) with the same name.
}
\keyword{internal}
