% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_util.R
\name{print_warn}
\alias{print_warn}
\title{Help function to print debugging messages in beyondWhittle(misc.R)}
\usage{
print_warn(msg)
}
\description{
Help function to print debugging messages in beyondWhittle(misc.R)
}
\keyword{internal}
