% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcWishart}
\alias{rcWishart}
\title{Generate random variables from the complex Wishart distribution}
\usage{
rcWishart(nu, Sigma_half)
}
\description{
Generate random variables from the complex Wishart distribution
}
\keyword{internal}
