% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sldmvnorm}
\alias{sldmvnorm}
\title{Sum of multivariate normal log densities
with mean 0 and covariance Sigma, unnormalized in beyondWhittle(varma.cpp)}
\usage{
sldmvnorm(z_t, Sigma)
}
\description{
Sum of multivariate normal log densities
with mean 0 and covariance Sigma, unnormalized in beyondWhittle(varma.cpp)
}
\keyword{internal}
