% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sqrt_cube}
\alias{sqrt_cube}
\title{Square root of a matrix array. See Remark 5.2}
\usage{
sqrt_cube(f_, excludeBoundary)
}
\description{
Square root of a matrix array. See Remark 5.2
}
\keyword{internal}
