% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{transfer_polynomial}
\alias{transfer_polynomial}
\title{VARMA transfer polynomials in beyondWhittle(varma.cpp)}
\usage{
transfer_polynomial(lambda, coef)
}
\description{
VARMA transfer polynomials in beyondWhittle(varma.cpp)
}
\keyword{internal}
