% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_multivariate_util.R
\name{uci_help}
\alias{uci_help}
\title{Helping function for \code{uci_matrix}. This is in beyondWhittle(misc.R).}
\usage{
uci_help(fpsd.sample, alpha, log = F)
}
\description{
Helping function for \code{uci_matrix}. This is in beyondWhittle(misc.R).
}
\keyword{internal}
