% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_multivariate_util.R
\name{uci_matrix}
\alias{uci_matrix}
\title{Uniform credible intervals in matrix-valued case. See (6.5)
This is in beyondWhittle(misc.R)}
\usage{
uci_matrix(fpsd.sample, alpha, uniform_among_components = F)
}
\description{
Uniform credible intervals in matrix-valued case. See (6.5)
This is in beyondWhittle(misc.R)
}
\keyword{internal}
