% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_multivariate_util.R
\name{uniformmax_multi}
\alias{uniformmax_multi}
\title{Helping function for \code{uci_matrix}
This is in beyondWhittle(misc.R).}
\usage{
uniformmax_multi(mSample)
}
\description{
Helping function for \code{uci_matrix}
This is in beyondWhittle(misc.R).
}
\keyword{internal}
